refactor: re-use `useEntityTypeEntitiesContextValue`
